Line decomp parallel (#459)
* IOSS: refactor to reduce strings; give flexibility and efficiency

* IOSS: Start of support for parallel line decomp

* IOSS: Explicit template instantiation fix

* IOSS: Allow passing filename down into line_decomp

* IOSS: initial line_decompose(); compile/link does not run

* IOSS: Refactor to try to eliminate duplicate code

* IOSS: Add some missing includes

* SLICE: Fix data_storage type

* Share decomp code between slice and ioss line decomp

* IOSS: Fix serial build

* Pull element centroid into common utils class

* Templative line_decompose; pass correct vector

* Handle specified in guided_decompose

* Remove unused function

* elementToProc does not need 64-bit range

* EXPLORE: Fix behavior after bad parse warning

* EXPLORE: Better warning/info message on SELECT

* Minor rearrange include files

* IOSS: See if this fixes/affects msys2 build

* IOSS: Add some logging/hwm code to line decomp

* IOSS: Another try to see how affects msys2 build

* IOSS: Better hwm logging output

* Unify Slice and DecompositionUtils zoltan_decompose

* IOSS: Fix msys2 build

* clang-format run

* SLICE: Version should be updated for latest chagnes

* IOSS: Reduce storage potentially; fix zoltan free call

* SLICE: Fix order of file close and mpi_finalize

* IOSS: Enable decomposition statistics for line decomp

* IOSS: io_shell - add_processor_id_field works for exodus also

* IOSS: compose output will add a proc_id map to output

* IOSS: thread-safe output_processor_id_map

* CI: safer variable naming

---------
