+++ packages/seacas/libraries/ioss/src/Ioss_ChainGenerator.C
@@ -84,14 +84,14 @@ namespace {
 
   template <typename INT>
   void get_line_front(Ioss::SideSet *fs, const Ioss::ElementBlock *block,
-                      const std::string &adj_block, Ioss::chain_t<INT> &element_chains,
-                      front_t<INT> &front)
+                      Ioss::chain_t<INT> &element_chains, front_t<INT> &front)
   {
+    const auto     adj_block_name = block->name();
     Ioss::NameList blocks;
     fs->block_membership(blocks);
     for (const auto &fs_block : blocks) {
-      if (fs_block == adj_block) {
-        // This faceset has some elements that are in `adj_block` -- put those in the `front`
+      if (fs_block == adj_block_name) {
+        // This faceset has some elements that are in `adj_block_name` -- put those in the `front`
         // list. Get list of "sides" in this faceset...
         std::vector<INT> element_side;
         assert(fs->side_block_count() == 1);
@@ -117,23 +117,19 @@ namespace {
   }
 
   template <typename INT>
-  front_t<INT> get_line_front(Ioss::Region &region, const std::string &adj_block,
+  front_t<INT> get_line_front(Ioss::Region &region, const Ioss::ElementBlock *block,
                               Ioss::chain_t<INT> &element_chains, const std::string &surface_list)
   {
     front_t<INT> front;
 
     // Since lines can not cross element blocks, we can process everything a block at a time.
-    const auto *block = region.get_element_block(adj_block);
     assert(block != nullptr);
-    if (block->topology()->shape() != Ioss::ElementShape::HEX) {
-      fmt::print("Skipping Element Block {}; it does not contain HEX elements.\n", adj_block);
-      return front;
-    }
+    assert(block->topology()->shape() == Ioss::ElementShape::HEX);
 
     if (surface_list == "ALL") {
       const Ioss::SideSetContainer &fss = region.get_sidesets();
       for (const auto &fs : fss) {
-        get_line_front(fs, block, adj_block, element_chains, front);
+        get_line_front(fs, block, element_chains, front);
       }
     }
     else {
@@ -141,7 +137,7 @@ namespace {
       for (const auto &surface : selected_surfaces) {
         auto *sset = region.get_sideset(surface);
         if (sset != nullptr) {
-          get_line_front(sset, block, adj_block, element_chains, front);
+          get_line_front(sset, block, element_chains, front);
         }
       }
     }
@@ -201,33 +197,47 @@ namespace Ioss {
     size_t             numel = region.get_property("element_count").get_int();
     Ioss::chain_t<INT> element_chains(numel);
 
-    // Generate the faces for use later...
-    Ioss::FaceGenerator face_generator(region);
-    face_generator.generate_faces((INT)0, true, true);
-
     // Determine which element block(s) are adjacent to the faceset specifying "lines"
     // The `adjacent_blocks` contains the names of all element blocks that are adjacent to the
     // surface(s) that specify the faces at the 'root' of the lines...
-    Ioss::NameList adjacent_blocks = get_adjacent_blocks(region, surface_list);
-    if (adjacent_blocks.empty()) {
+    Ioss::NameList adjacent_block_names = get_adjacent_blocks(region, surface_list);
+    if (adjacent_block_names.empty()) {
       fmt::print("WARNING: No surfaces in the model matched the input surface list ({}).\n\tNo "
                  "chains will be generated.\n",
                  surface_list);
     }
-    for (const auto &adj_block : adjacent_blocks) {
-      // Get the offset into the element_chains vector...
-      const auto *block  = region.get_element_block(adj_block);
-      auto        offset = block->get_offset() + 1;
-      auto        count  = block->entity_count();
 
-      auto front = get_line_front(region, adj_block, element_chains, surface_list);
+    // Get the EB* corresponding to the EB names...
+    Ioss::ElementBlockContainer adjacent_blocks;
+    adjacent_blocks.reserve(adjacent_block_names.size());
+    for (const auto &blk_name : adjacent_block_names) {
+      auto *eb = region.get_element_block(blk_name);
+      assert(eb != nullptr);
+      if (eb->topology()->shape() != Ioss::ElementShape::HEX) {
+        fmt::print("Skipping Element Block {}; it does not contain HEX elements.\n", blk_name);
+      }
+      else {
+        adjacent_blocks.push_back(eb);
+      }
+    }
+
+    // Generate the faces for use later... (only generate on the blocks touching the front)
+    Ioss::FaceGenerator face_generator(region);
+    face_generator.generate_block_faces(adjacent_blocks, (INT)0, true);
+
+    for (const auto *block : adjacent_blocks) {
+      // Get the offset into the element_chains vector...
+      auto offset = block->get_offset() + 1;
+      auto count  = block->entity_count();
+
+      auto front = get_line_front(region, block, element_chains, surface_list);
       if (front.empty()) {
         continue;
       }
 
       // We want a vector giving us the Face for each face of each element in the block...
       connectivity_t face_connectivity(count);
-      generate_face_connectivity(face_generator.faces(adj_block), static_cast<int>(offset),
+      generate_face_connectivity(face_generator.faces(block), static_cast<int>(offset),
                                  face_connectivity);
 
       // For each face on the "front" (at the beginning the boundary sideset faces)
